backend-chat/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
openai
groq